+++ MVC/Services/IJsonDbServices.cs
@@ -5,7 +5,7 @@ namespace MVC.Services
     public interface IJsonDbServices
     {
         public List<UserReview> GetUserReviewToJSON();
-        public List<UserReview> GetUserReviewToJSONById(int id);
+        public List<UserReview> GetUserReviewToJSONById(long id);
         bool AddNewUserReviewToJSON(UserReview review);
         bool DeleteUserReviewToJSON(UserReview review);
         bool EditUserReviewToJSON(UserReview review);
